Time: 1876 ms (5.04%), Space: 7.7 MB (87.23%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     bool isPerfectSquare(int num) {
         if (num==1) return true;
-        for(int i =1;i<num;i++){
-            long x=(long)i*i;
+        for(long  i =1;i<num;i++){
+            long x=i*i;
             if(x==num){
                 return true;
             }
